[Refactor] Remove unused typedefs
When starting the flag_set, I was thinking of it making it a true
container and started to define some sensible typedefs, but did not
continue with defining the iterator typedefs needed. So, for now,
they are just noise and in the way. So, we remove them.

At a later point, we may define the full typedefs needed to be
a container.